--- src/pages/PlaceFilter/PlaceFilter.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import PlaceItem from '../../components/PlaceItem';
-
-const PlaceHomeFilter = () => {
-  return (
-    <div>
-      <PlaceItem />
-      <PlaceItem />
-    </div>
-  );
-};
-
-export default PlaceHomeFilter;
